requirements.txt
Django>=4.2
django-ratelimit==3.0.1
